Correctly call the tailwindcss preset if the auth flag is not passed
Due to the removal of `tailwindcss-auth` macro the `tailwindcss` macro
will always run `TailwindCssPreset::installAuth()`. This is in large
part because how macro's work and the `UiCommand::macro('tailwindcss')`
getting overwritten by `AuthCommand::macro('tailwindcss')`.

This commit assumes that just having one macro of `tailwindcss` is the
acceptable use moving forward. With that said, this commit check the
command option auth flag and if it's true will correctly install all the
auth presets.

This commit also removes the call to the `install()` method on the
TailwindCssPreset::installAuth method as this should always be called on
the macro.